+++ server/src/weaverbird/pipeline/pipeline.py
@@ -257,11 +257,10 @@ def _remove_void_from_condition(condition: Condition) -> Condition | None:
         if condition.column == VOID_REPR or condition.value == VOID_REPR:
             return None
     elif isinstance(condition, InclusionCondition):
-        condition_values = [v for v in condition.value if v != VOID_REPR]
-        if len(condition_values) == 0:
+        if condition.value == VOID_REPR:
             return None
-        condition.value = condition_values
-
+        condition_values = [v for v in condition.value if v != VOID_REPR]
+        return condition.model_copy(update={"value": condition_values})
     return condition