+++ server/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixed
 
-- The `remove_void_conditions_from_filter_steps` function does not remove anymore `InclusionCondition` with an empty list of values to match. 
+- The `remove_void_conditions_from_filter_steps` function does not remove `InclusionCondition` with an empty list of values anymore. 
 
 ## [0.50.1] - 2025-02-04